html/semantics/forms/event-propagate-disabled.html

<!DOCTYPE html>
<title>Event propagation on disabled form elements</title>
<meta charset=utf8>
<meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<div id=cases>
  <input> <!-- Santity check with non-disabled control -->
  <select disabled></select>
  <select disabled>
    <option>foo</option>
  </select>
  <fieldset disabled>Text</fieldset>
  <fieldset disabled><span>Span</span></fieldset>
  <button disabled>Text</button>
  <button disabled><span>Span</span></button>
  <textarea disabled></textarea>
  <input disabled type="button">
  <input disabled type="checkbox">
  <input disabled type="color" value="#000000">
  <input disabled type="date">
  <input disabled type="datetime-local">
  <input disabled type="email">
  <input disabled type="image">
  <input disabled type="month">
  <input disabled type="number">
  <input disabled type="password">
  <input disabled type="radio">
  <!-- Native click will click the bar -->
  <input disabled type="range" value="0">
  <!-- Native click will click the slider -->
  <input disabled type="range" value="50">
  <input disabled type="reset">
  <input disabled type="search">
  <input disabled type="submit">
  <input disabled type="tel">
  <input disabled type="text">
  <input disabled type="time">
  <input disabled type="url">
  <input disabled type="week">

  <!--
    WebDriver can't click on type="file"
    https://github.com/w3c/webdriver/issues/1666
  <input disabled="disabled" type="file">
  <input disabled="disabled" type="hidden"> -->
</div>

<script>
  /**
   * @param {Element} element
   */
  function getEventFiringTarget(element) {
    let target = element;
    while (target.children.length) {
      const { firstElementChild } = target;
      if (!firstElementChild.getBoundingClientRect().width) {
        // Some elements does not have its own painting area, e.g. <option>.
        // Such element can't be clicked natively.
        break;
      }
      target = firstElementChild;
    }
    return target;
  }

  const allEvents = ["pointermove", "mousemove", "pointerdown", "mousedown", "pointerup", "mouseup", "click"];

  /**
   * @param {*} t
   * @param {Element} element
   * @param {Element} observingElement
   */
  function setupTest(t, element, observingElement) {
    /** @type {string[]} */
    const observedEvents = [];
    const controller = new AbortController();
    const { signal } = controller;
    const listenerFn = t.step_func(event => observedEvents.push(event.type));
    for (const event of allEvents) {
      observingElement.addEventListener(event, listenerFn, { signal });
    }
    t.add_cleanup(() => controller.abort());

    const target = getEventFiringTarget(element);
    return { target, observedEvents };
  }

  /**
   * @param {object} options
   * @param {Element & { disabled: boolean }} options.element
   * @param {string[]} options.expectedEvents
   * @param {(target: Element) => (Promise<void> | void)} options.clickerFn
   * @param {string} options.title
   */
  function promise_event_test({ element, expectedEvents, nonDisabledExpectedEvents, clickerFn, title }) {
    for (const observingElement of [element, document.body]) {
      promise_test(async t => {
        const { target, observedEvents } = setupTest(t, element, observingElement);

        await clickerFn(target);
        await new Promise(resolve => t.step_timeout(resolve, 0));

        const expected = element.disabled ? expectedEvents : nonDisabledExpectedEvents;
        assert_array_equals(observedEvents, expected, "Observed events");

      }, `${title} on ${element.outerHTML}, observed from <${observingElement.localName}>`);
    }
  }

  const mouseEvents = ["mousemove", "mousedown", "mouseup", "click"];
  const pointerEvents = ["pointermove", "pointerdown", "pointerup"];

  const elements = document.getElementById("cases").children;
  for (const element of elements) {
    promise_event_test({
      element,
      expectedEvents: pointerEvents,
      nonDisabledExpectedEvents: allEvents,
      clickerFn: test_driver.click,
      title: "Trusted click"
    })

    const eventFirePair = [
      [MouseEvent, mouseEvents],
      [PointerEvent, pointerEvents]
    ];

    for (const [eventInterface, events] of eventFirePair) {
      promise_event_test({
        element,
        expectedEvents: events,
        nonDisabledExpectedEvents: events,
        clickerFn: target => {
          for (const event of events) {
            target.dispatchEvent(new eventInterface(event, { bubbles: true }))
          }
        },
        title: `Dispatch new ${eventInterface.name}()`
      })
    }

    promise_event_test({
      element,
      expectedEvents: [],
      nonDisabledExpectedEvents: ["click"],
      clickerFn: target => target.click(),
      title: `click()`
    })
  }
</script>
